Put modules into separate subdirectories
