+++ src/core/GuildSettings.cpp
@@ -1,5 +1,5 @@
 #include "GuildSettings.h"
-#include "modules/LevelModule.h"
+#include "modules/level/LevelModule.h"
 
 #include "Logger.h"
