+++ src/modules/currency/CurrencyModule.h
