+++ src/modules/level/LevelModule.h
@@ -47,4 +47,6 @@ private:
 	QTimer m_timer;
 
 	UmikoBot* m_client;
+
+	void initiate_admin_commands();
 };
