+++ src/modules/timezone/TimezoneModule.h
@@ -1,4 +1,5 @@
 #pragma once
+
 #include <QtCore/QTimeZone>
 #include "core/Module.h"
